--- .gitignore
@@ -1,4 +1,0 @@
-alwaysai.target.json
-.vscode/
-.DS_Store
-__pycache__
