--- Dockerfile
@@ -1,1 +1,0 @@
-FROM alwaysai/edgeiq:0.9.0
