--- alwaysai.app.json
@@ -1,8 +1,0 @@
-{
-  "models": {
-    "alwaysai/mobilenet_ssd": 1
-  },
-  "scripts": {
-    "start": "python app.py"
-  }
-}
