+++ docker-compose.yml
@@ -1,0 +1,42 @@
+version: '2'
+volumes:
+  sense-data:
+services:
+  influxdb:
+    restart: always
+    build: ./balena-sense/influxdb
+    volumes:
+      - 'sense-data:/data'
+  grafana:
+    restart: always
+    build: ./balena-sense/grafana
+    ports:
+      - "80"
+    volumes:
+      - 'sense-data:/data'
+    environment:
+        - 'GF_SERVER_HTTP_PORT=80'
+        - 'GF_PATHS_PROVISIONING=/usr/src/app/provisioning'
+        - 'GF_SESSION_PROVIDER=memory'
+  sensor:
+    build: ./balena-sense/sensor
+    privileged: true
+    restart: always
+    volumes:
+      - 'sense-data:/data'
+  telegraf:
+    build: ./balena-sense/telegraf
+    restart: always
+    cap_add:
+      - SYS_ADMIN
+  mqtt:
+    image: arm32v6/eclipse-mosquitto
+    ports:
+      - "1883:1883"
+    restart: always
+  people-counter:
+    build: ./people-counter
+    privileged: true
+    restart: always
+    ports:
+      - 5000:5000
