+++ people-counter/.gitignore
@@ -1,0 +1,5 @@
+alwaysai.target.json
+.vscode/
+.DS_Store
+__pycache__
+*.tmp
