+++ people-counter/alwaysai.app.json
@@ -1,0 +1,8 @@
+{
+  "models": {
+    "alwaysai/mobilenet_ssd": 1
+  },
+  "scripts": {
+    "start": "python app.py"
+  }
+}
