+++ people-counter/models/alwaysai/mobilenet_ssd/alwaysai.model.json
@@ -1,0 +1,31 @@
+{
+  "id": "alwaysai/mobilenet_ssd",
+  "accuracy": "",
+  "description": "Classifies 20 PASCAL object classes: person, bird, cat, cow, dog, horse, sheep, aeroplane, bicycle, boat, bus, car, motorbike, train, bottle, chair, dining table, potted plant, sofa, tv/monitor.",
+  "inference_time": 0.18125,
+  "license": "MIT",
+  "mean_average_precision_top_1": 0.727,
+  "mean_average_precision_top_5": null,
+  "public": true,
+  "dataset": "COCO",
+  "website_url": "https://github.com/chuanqi305/MobileNet-SSD",
+  "model_parameters": {
+    "framework_type": "caffe",
+    "config_file": "models/MobileNetSSD_deploy.prototxt.txt",
+    "size": [
+      300,
+      300
+    ],
+    "model_file": "models/MobileNetSSD_deploy.caffemodel",
+    "label_file": "labels/pascal_voc.txt",
+    "scalefactor": 0.007843,
+    "mean": [
+      127.5,
+      127.5,
+      127.5
+    ],
+    "crop": true,
+    "purpose": "ObjectDetection"
+  },
+  "version": 1
+}
